requirements.txt
Django ~= 3.1
djangorestframework == 3.12.4
psycopg2 == 2.9.1
python-dotenv == 0.18.0
djangorestframework-simplejwt == 4.7.2
django-filter == 2.4.0
drf-yasg ~= 1.20.0
